refactor(modulos): F3 cirúrgica PR-2 — análise SPIN é telefonia (−13 arestas, quebra o ciclo farmer↔telefonia)
Move a feature SPIN (análise de chamada / copilot da ligação) de farmer-inteligencia
para telefonia-whatsapp-rota: git mv src/lib/spin + src/hooks/useSpinAnalysis[.test]
→ src/lib/call/spin/ (co-locada; destino já capturado pelo glob src/lib/call/** da telefonia).

Todos os consumidores runtime são da telefonia (WebRTCCallContext, TranscriptionPanel,
SpinSuggestionCard, lib/call-session/*); único leitor externo é admin-crm/CallSessionDetail
(re-rotulado admin-crm→telefonia). Zero consumo no farmer além do próprio hook.

Manifesto: removidas as 3 entradas mortas do farmer (lib/spin/**, useSpinAnalysis.ts,
o teste); zero adição. Baseline 239→226 (−13): 11 telefonia→farmer + 2 farmer→telefonia
(lado-origem useSpinAnalysis[.test]→lib/transcription) — o ciclo bidirecional encolhe dos dois lados.

Provas: manifesto.gate + fronteiras.gate (67 testes modulos), typecheck 0 erros,
lint 0 erros, suíte global 584 arquivos/4746 testes — todas verdes.

Colisão sinalizada: PR #1212 (deadcode) edita lib/spin/types.ts in-place → quem mergear em 2º re-aplica no novo caminho.
