+++ docs/historico/modularizacao.md
@@ -6,8 +6,9 @@
 
 - **Re-decisão (founder + Codex `gpt-5.5` reasoning high):** o parecer de 2026-07-08 se confirmou e reforçou — *"o problema restante é fronteira lógica, não diretório; mover tudo para `src/modules/` não reduz uma única aresta, só troca conflito de arquitetura por conflito de Git"*. Contexto que pesou: PR #1212 (~100 arquivos de src/) parado + 8 sessões paralelas vivas. Founder aprovou o cirúrgico guiado pelo relatório de fronteiras.
 - **Critério Codex para burn-down (vale para os próximos):** aresta **legítima** = capacidade estável expressa como verbo, sem importar UI/hook/contexto interno do outro módulo; **vazamento** = importa interno alheio, cria bidirecionalidade ou fan-out. Par bidirecional com dezenas de arestas = módulo mal cortado ou contrato ausente. ⚠️ Anti-padrão nomeado: plataforma como **"lixão de contratos"** — contrato em plataforma só com linguagem transversal (aparecer `Farmer`/`SPIN`/`RoutePlanner` no nome = acoplamento lavado).
-- **PR-1 — routePlanner → telefonia (este PR).** As 45 arestas do par telefonia↔reposicao (16% da baseline F2) eram **erro de corte, não acoplamento**: `components/reposicao/routePlanner/` tinha todos os consumidores (`AdminRoutePlanner`, `useRoutePlanner`, `lib/route/*`) e todas as dependências (`lib/route`, `lib/maps`, `BotaoLigar`) na telefonia; zero consumo pela reposição. `git mv` → `components/rota/planner/` (24 arquivos, histórico preservado) + import-fix em 17 consumidores. **Manifesto intacto** (glob `components/rota/**` da telefonia já capturava o destino — gate F1 confirmou sem órfão/sobreposição). Baseline **284→239** (−45 líquido; 6 arestas re-rotuladas honestamente p/ farmer↔telefonia, alvo do PR-2). Método que achou o alvo: detalhar a baseline por par (`grep deModulo/paraModulo`) — o fan-out 100% concentrado numa pasta denunciou o corte errado.
-- **Próximos:** PR-2 contrato SPIN (11 das 17 arestas telefonia→farmer; quebra o ciclo — tem design a fazer, cuidado com o anti-padrão acima); PR-3 sino do AppShell (7 inversões plataforma→negócio + ~30 KB eager, já quantificado na F4).
+- **PR-1 — routePlanner → telefonia (#1262).** As 45 arestas do par telefonia↔reposicao (16% da baseline F2) eram **erro de corte, não acoplamento**: `components/reposicao/routePlanner/` tinha todos os consumidores (`AdminRoutePlanner`, `useRoutePlanner`, `lib/route/*`) e todas as dependências (`lib/route`, `lib/maps`, `BotaoLigar`) na telefonia; zero consumo pela reposição. `git mv` → `components/rota/planner/` (24 arquivos, histórico preservado) + import-fix em 17 consumidores. **Manifesto intacto** (glob `components/rota/**` da telefonia já capturava o destino — gate F1 confirmou sem órfão/sobreposição). Baseline **284→239** (−45 líquido; 6 arestas re-rotuladas honestamente p/ farmer↔telefonia, alvo do PR-2). Método que achou o alvo: detalhar a baseline por par (`grep deModulo/paraModulo`) — o fan-out 100% concentrado numa pasta denunciou o corte errado.
+- **PR-2 — análise SPIN → telefonia (este PR).** SPIN é análise de **chamada** (copilot da ligação), não do farmer: **todos** os consumidores runtime são telefonia (`WebRTCCallContext`, `TranscriptionPanel`, `SpinSuggestionCard`, `lib/call-session/*`) + 1 leitor admin-crm (`CallSessionDetail`, re-rotulado 1:1); **zero** consumo no farmer além do próprio hook. `git mv src/lib/spin` + `useSpinAnalysis[.test]` → **`src/lib/call/spin/`** (feature co-locada: types+prompts+hook+testes juntos; destino já capturado pelo glob `src/lib/call/**` — mesma elegância "manifesto intacto no destino" do PR-1) + import-fix em 13 arquivos. Manifesto: removidas as **3 entradas mortas** do farmer (`lib/spin/**`, `useSpinAnalysis.ts`, o teste); **zero adição** (gate `manifesto.gate` verde: 1-dono/arquivo, sem órfão/glob-morto). Baseline **239→226 (−13)**: 11 `telefonia→farmer` (as previstas pelo PR-1) **+ 2 `farmer→telefonia`** (lado-origem `useSpinAnalysis[.test]→lib/transcription`, agora intra-telefonia) — o ciclo encolhe **dos dois lados**; 1 aresta admin-crm re-rotulada honestamente (`admin-crm→telefonia`, leitor legítimo do contrato). ⚠️ Sem "lixão de contratos": `SpinAnalysis` tem **nome de domínio** → dono telefonia, não plataforma (`types.ts` = contrato puro, 92 linhas, zero imports). Convenção: 1º hook sob `src/lib/` — troca deliberada (coesão física da feature, goal do F3). Colisão sinalizada: PR #1212 (deadcode) edita `lib/spin/types.ts` in-place → quem mergear em 2º re-aplica no novo caminho.
+- **Próximos:** PR-3 sino do AppShell (7 inversões plataforma→negócio + ~30 KB eager, já quantificado na F4).
 
 ## ✅ F2 — gate de fronteiras anti-vazamento com ratchet (2026-07-08, PR #1255)
