+++ src/components/call/TranscriptionPanel.tsx
@@ -5,7 +5,7 @@ import { cn } from '@/lib/utils';
 import { Button } from '@/components/ui/button';
 import { Badge } from '@/components/ui/badge';
 import type { TranscriptTurn, TranscriptionStatus } from '@/lib/transcription/types';
-import type { SpinAnalysis, SpinAnalysisStatus } from '@/lib/spin/types';
+import type { SpinAnalysis, SpinAnalysisStatus } from '@/lib/call/spin/types';
 import { SpinSuggestionCard } from './SpinSuggestionCard';
 
 interface TranscriptionPanelProps {
