+++ src/contexts/WebRTCCallContext.tsx
@@ -8,8 +8,8 @@ import { normalizeBrPhone, formatBrPhone } from '@/lib/phone';
 import { mixPrerollWithMic } from '@/lib/sip/audio-preroll';
 import { useTranscription } from '@/hooks/useTranscription';
 import type { TranscriptTurn } from '@/lib/transcription/types';
-import { useSpinAnalysis } from '@/hooks/useSpinAnalysis';
-import type { SpinAnalysis } from '@/lib/spin/types';
+import { useSpinAnalysis } from '@/lib/call/spin/useSpinAnalysis';
+import type { SpinAnalysis } from '@/lib/call/spin/types';
 import { resolveCustomerByPhone } from '@/lib/call-session/resolve-customer';
 import { buildSessionPayload } from '@/lib/call-session/build-session-payload';
 import { resolveCallParty, shouldAutoRecord, type ResolvedCallParty } from '@/lib/call-log/recording-policy';
