+++ src/lib/modulos/fronteiras-baseline.ts
@@ -30,10 +30,8 @@ export const FRONTEIRAS_BASELINE: Aresta[] = [
   { de: "src/components/caca/FilaDeCaca.tsx", para: "src/components/call/BotaoLigar.tsx", deModulo: "caca", paraModulo: "telefonia-whatsapp-rota", kind: "runtime" },
   { de: "src/components/call/FichaPreContato.tsx", para: "src/hooks/useHistoricoCompras.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "vendas", kind: "runtime" },
   { de: "src/components/call/OutcomeMenu.tsx", para: "src/queries/useRegistrarContato.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "runtime" },
-  { de: "src/components/call/SpinSuggestionCard.tsx", para: "src/lib/spin/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "type" },
-  { de: "src/components/call/TranscriptionPanel.tsx", para: "src/lib/spin/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "type" },
   { de: "src/components/customer/CallSessionDetail.tsx", para: "src/hooks/useCustomerCalls.ts", deModulo: "admin-crm", paraModulo: "telefonia-whatsapp-rota", kind: "type" },
-  { de: "src/components/customer/CallSessionDetail.tsx", para: "src/lib/spin/types.ts", deModulo: "admin-crm", paraModulo: "farmer-inteligencia", kind: "type" },
+  { de: "src/components/customer/CallSessionDetail.tsx", para: "src/lib/call/spin/types.ts", deModulo: "admin-crm", paraModulo: "telefonia-whatsapp-rota", kind: "type" },
   { de: "src/components/customer/CallSessionRow.tsx", para: "src/hooks/useCustomerCalls.ts", deModulo: "admin-crm", paraModulo: "telefonia-whatsapp-rota", kind: "type" },
   { de: "src/components/customer/CustomerBrief.tsx", para: "src/components/customer360/format.ts", deModulo: "admin-crm", paraModulo: "farmer-inteligencia", kind: "runtime" },
   { de: "src/components/customer/CustomerBrief.tsx", para: "src/hooks/useCustomerCalls.ts", deModulo: "admin-crm", paraModulo: "telefonia-whatsapp-rota", kind: "runtime" },
@@ -120,15 +118,10 @@ export const FRONTEIRAS_BASELINE: Aresta[] = [
   { de: "src/components/unified-order/TierClienteBadge.tsx", para: "src/hooks/useClienteTier.ts", deModulo: "vendas", paraModulo: "farmer-inteligencia", kind: "runtime" },
   { de: "src/components/visitas/AgendarVisitaDialog.tsx", para: "src/components/customer/CustomerBrief.tsx", deModulo: "farmer-inteligencia", paraModulo: "admin-crm", kind: "runtime" },
   { de: "src/contexts/__tests__/DashboardPersonaContext.test.tsx", para: "src/lib/dashboard/route-tracker.ts", deModulo: "plataforma", paraModulo: "farmer-inteligencia", kind: "runtime" },
-  { de: "src/contexts/__tests__/WebRTCCallContext.test.tsx", para: "src/hooks/useSpinAnalysis.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "runtime" },
   { de: "src/contexts/DashboardPersonaContext.tsx", para: "src/lib/dashboard/persona-config.ts", deModulo: "plataforma", paraModulo: "farmer-inteligencia", kind: "type" },
   { de: "src/contexts/DashboardPersonaContext.tsx", para: "src/lib/dashboard/persona-detect.ts", deModulo: "plataforma", paraModulo: "farmer-inteligencia", kind: "type" },
-  { de: "src/contexts/webrtc-call-context.ts", para: "src/lib/spin/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "type" },
-  { de: "src/contexts/WebRTCCallContext.tsx", para: "src/hooks/useSpinAnalysis.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "runtime" },
-  { de: "src/contexts/WebRTCCallContext.tsx", para: "src/lib/spin/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "type" },
   { de: "src/hooks/__tests__/lens-engines-ia.test.tsx", para: "src/hooks/useCrossSellEngine.ts", deModulo: "farmer-inteligencia", paraModulo: "vendas", kind: "runtime" },
   { de: "src/hooks/__tests__/offline-flow-integration.test.tsx", para: "src/services/picking-confirm.ts", deModulo: "plataforma", paraModulo: "estoque-recebimento", kind: "type" },
-  { de: "src/hooks/__tests__/useSpinAnalysis.test.tsx", para: "src/lib/transcription/types.ts", deModulo: "farmer-inteligencia", paraModulo: "telefonia-whatsapp-rota", kind: "type" },
   { de: "src/hooks/dashboard/useFinanceiroZone.ts", para: "src/services/financeiroService.ts", deModulo: "farmer-inteligencia", paraModulo: "financeiro", kind: "runtime" },
   { de: "src/hooks/unifiedOrder/useCoresDoCliente.ts", para: "src/lib/tint/cores-do-cliente.ts", deModulo: "vendas", paraModulo: "tintometrico", kind: "runtime" },
   { de: "src/hooks/useAdminOrderDetail.ts", para: "src/hooks/usePriceHistory.ts", deModulo: "admin-crm", paraModulo: "vendas", kind: "runtime" },
@@ -159,7 +152,6 @@ export const FRONTEIRAS_BASELINE: Aresta[] = [
   { de: "src/hooks/useProcessComparison.ts", para: "src/lib/customer-process/comparison-types.ts", deModulo: "knowledge-base", paraModulo: "admin-crm", kind: "type" },
   { de: "src/hooks/useRoutePlanner.ts", para: "src/hooks/useFarmerScoring.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "runtime" },
   { de: "src/hooks/useRoutePlanner.ts", para: "src/lib/visitas/agenda-to-stop.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "runtime" },
-  { de: "src/hooks/useSpinAnalysis.ts", para: "src/lib/transcription/types.ts", deModulo: "farmer-inteligencia", paraModulo: "telefonia-whatsapp-rota", kind: "type" },
   { de: "src/hooks/useUnifiedOrder.ts", para: "src/hooks/useClienteTier.ts", deModulo: "vendas", paraModulo: "farmer-inteligencia", kind: "runtime" },
   { de: "src/hooks/useUnifiedOrder.ts", para: "src/hooks/useRecommendationEngine.ts", deModulo: "vendas", paraModulo: "farmer-inteligencia", kind: "type" },
   { de: "src/hooks/useWhatsappPendentes.ts", para: "src/lib/fila/adapters/whatsappPendente.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "type" },
@@ -167,11 +159,6 @@ export const FRONTEIRAS_BASELINE: Aresta[] = [
   { de: "src/lib/__tests__/offline-handlers.test.ts", para: "src/services/recebimento-confirm.ts", deModulo: "plataforma", paraModulo: "estoque-recebimento", kind: "runtime" },
   { de: "src/lib/__tests__/offline-handlers.test.ts", para: "src/services/recebimento-cte.ts", deModulo: "plataforma", paraModulo: "estoque-recebimento", kind: "runtime" },
   { de: "src/lib/__tests__/offline-handlers.test.ts", para: "src/services/recebimento-divergencia.ts", deModulo: "plataforma", paraModulo: "estoque-recebimento", kind: "runtime" },
-  { de: "src/lib/call-session/__tests__/build-session-payload.test.ts", para: "src/lib/spin/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "type" },
-  { de: "src/lib/call-session/aggregate-entities.test.ts", para: "src/lib/spin/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "type" },
-  { de: "src/lib/call-session/aggregate-entities.ts", para: "src/lib/spin/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "type" },
-  { de: "src/lib/call-session/build-session-payload.test.ts", para: "src/lib/spin/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "type" },
-  { de: "src/lib/call-session/build-session-payload.ts", para: "src/lib/spin/types.ts", deModulo: "telefonia-whatsapp-rota", paraModulo: "farmer-inteligencia", kind: "type" },
   { de: "src/lib/carteira/escopo-clientes.ts", para: "src/components/adminCustomers/types.ts", deModulo: "farmer-inteligencia", paraModulo: "admin-crm", kind: "type" },
   { de: "src/lib/fila/__tests__/rota.test.ts", para: "src/queries/useRouteContactList.ts", deModulo: "farmer-inteligencia", paraModulo: "telefonia-whatsapp-rota", kind: "type" },
   { de: "src/lib/fila/__tests__/tarefa.test.ts", para: "src/lib/tarefas/types.ts", deModulo: "farmer-inteligencia", paraModulo: "tarefas", kind: "type" },
