+++ src/lib/modulos/manifesto.ts
@@ -193,7 +193,6 @@ export const MODULOS: ModuloApp[] = [
       "src/lib/visit-scoring/**",
       "src/lib/dashboard/**",
       "src/lib/sinais/**",
-      "src/lib/spin/**",
       "src/lib/fila/**",
       "src/components/fila/**",
       "src/components/farmer/**",
@@ -228,7 +227,6 @@ export const MODULOS: ModuloApp[] = [
       "src/hooks/useTeamKpis.ts",
       "src/hooks/useTeamRanking.ts",
       "src/hooks/useCoverage.ts",
-      "src/hooks/useSpinAnalysis.ts",
       "src/hooks/useBundleArguments.ts",
       "src/hooks/useBundleEngine.ts",
       "src/hooks/useCopilotEngine.ts",
@@ -260,7 +258,6 @@ export const MODULOS: ModuloApp[] = [
       "src/hooks/__tests__/lens-tactical-plan.test.tsx",
       "src/hooks/__tests__/tactical-plan-posse-dono.test.tsx",
       "src/hooks/__tests__/useLastVisit.test.tsx",
-      "src/hooks/__tests__/useSpinAnalysis.test.tsx",
       "src/hooks/__tests__/useDashboardCompany.test.ts",
     ],
     risco: { moneyPath: false, offlineFirst: false, authSensitive: false },
